requirements.txt
langchain
python-dotenv
langchain-openai
ipykernel
streamlit
faiss-cpu
langchain-community